Add option to gaussian filtering average FF
Add option to apply a gaussian filtering in order to blur the FF
average. This allows to remove the diffraction pattern observed
sometimes when averaging the FF images.

The std used at this moment is 16 pixels (radius of the filter).
The default option is to not apply the filter -g=0. In order to
apply it use the option -g=1.