Add option to change radius of gaussin filter
Add option to change standard deviation (std), which is related
to the radius of blurring of the gaussian filter.

By default -g=0 and the gaussian filter is not applied. Otherwise,
if an integer different than 0 is used, this integer indicates
the std/radius of the filter. Bigger it is, higher it is the
blurring effect of the filter.