Do not flip FF images
To maintain the correspondence between FF images in tomo hdf5 and
in mosaics hdf5 it seemed that the FF stored in mosaics had to
be flipped. Thus a np.flip was applied in mosaicnex.py.
But the mosaics were not well normalized applying this flip.
Remove the np.flip from the FF in order to solve the bug in
normalized mosaic hdf5 files.