Remove 'region' from xrm filename pattern
Accorded with the scientists of Mistral, that the new pattern that
will be used for xrm files will not contain the 'region' field.
Remove it.

The new pattern to be followed must have the following structure:
date_tomoname_energy_angle_ZP.xrm

Ex:
20161126_tomo05_520_10.0_-11350.0.xrm

Adapt code to deal with this new pattern.