Add skeleton for magnify script
Add skeleton for magnify script.

This script will be used at BL09-Mistral from ALBA synchrotron in
order to magnify by different ratios the normalized images of a
spectroscopy stack of images.

This can be used when problems in the movement of the CCD exit,
and the images cannot be magnified correctly using HW. Then, we
will be able to use magnify to apply a SW magnification.

An input file of magnification ratios will be the second input
of this script; the first being the spectroscopy normalized image
stack.