Begin to add workflow for xtendof
Begin to add a workflow for xtend_1; named for now xtendof, to
do not have conflicts with the old version of xtend_1 working
with stacks from almost the beginning.

This version should work with simple images till almost the end.
After the averaging (fusion), the images are merged in a stack.