Allow parsing jj positions
Allow parsing jj positions. This is useful in
magnetism experiments, for instance; where the light polarization
has to be choosen, and this can be done by setting the JJ offset
to different positions. With changes in the jj positions,
we can find the circular left and right polarizations.

This new parsing, will allow to preprocess data collected for
magnetism, many repetitions per angle, experiments.