Magnetism jj pipeline allow --db --ff many times
For the online workflow (partial preprocessing) a bug existed
if processing multiple times the ffs for the magnetism
(jj based polarization experiments) pipeline.

The FF ..._proc.hdf5 (processed) images where added many times
in the tinyDB database, and because of that the processing was not
done correclty, if for some reason, the 'magnetism --db --ff'
was applied more than once. Fix it.

It also fixes problem in case of applying more than once,
the full pipeline offline:
magnetism <FILE_NAME>.txt --db --ff --th --stack.
If it was applied more than once, the workflow was crashing,
if not removing the index.json and intermediate _proc.hdf5 files.
This commit fixes it as well.