Magnetism jj pipeline allow --th angle many times
For the online workflow (partial preprocessing) a bug existed
if processing multiple times a given angle for the magnetism
(jj based polarization experiments) pipeline.

The angle theta ..._proc.hdf5 (processed) images where added many times
in the tinyDB database, and because of that the processing was not
done correclty, if for some reason, the 'magnetism --th <value>'
was applied more than once. For example, to recalculate the images
of a given angle, if they were not correctly calculated.
Fix it.